+++ src/controllers/genreController.ts
@@ -7,18 +7,19 @@ import Genre, { IGenreAttributes } from "../models/genreModel";
 import Category from "../models/categoryModel";
 import Application from "../models/applicationModel";
 import { getGenreIdByName } from "../helper/genreHelper";
-const createGenreController = async (
-  req: Request,
-  res: Response
-): Promise<void> => {
+const createGenre = async (req: Request, res: Response): Promise<void> => {
   try {
-    const genre: string = req.body.genre;
-    const category: string = req.body.category;
+    let genre: string = req.body.genre;
+    let category: string = req.body.category;
+
+    genre = genre.trim();
+    category = category.trim();
 
     const categoryId: number | null = await getCategoryIdByName(category);
 
     if (!categoryId) {
       res.json({ success: 0, error: "Wrong category" });
+      return;
     }
     const newGenre: IGenreAttributes = {
       genre: genre,
@@ -31,10 +32,7 @@ const createGenreController = async (
   }
 };
 
-const getAllGenreController = async (
-  req: Request,
-  res: Response
-): Promise<void> => {
+const getAllGenre = async (req: Request, res: Response): Promise<void> => {
   try {
     const allGenres: Genre[] = await db.Genre.findAll({
       raw: true,
@@ -57,15 +55,26 @@ const getAllGenreController = async (
   }
 };
 
-const editGenreByIdController = async (
-  req: Request,
-  res: Response
-): Promise<void> => {
+const editGenreById = async (req: Request, res: Response): Promise<void> => {
   try {
-    const genre: string = req.body.genre;
-    const category: string = req.body.category;
+    let genre: string = req.body.genre;
+    let category: string = req.body.category;
     const id: string = req.body.id;
 
+    genre = genre.trim();
+    category = category.trim();
+
+    const findRes: Genre | null = await db.Genre.findOne({
+      where: {
+        id: id,
+      },
+      attributes: ["id"],
+    });
+
+    if (!findRes) {
+      res.json({ success: 0, error: "No genre to edit" });
+      return;
+    }
     const categoryId: Category | null = await db.Category.findOne({
       attributes: ["id"],
       raw: true,
@@ -95,7 +104,17 @@ const editGenreByIdController = async (
 const deleteGenereById = async (req: Request, res: Response): Promise<void> => {
   try {
     const id: string = req.body.id;
+    const findRes: Genre | null = await db.Genre.findOne({
+      where: {
+        id: id,
+      },
+      attributes: ["id"],
+    });
 
+    if (!findRes) {
+      res.json({ success: 0, error: "No genre to delete" });
+      return;
+    }
     await db.Genre.destroy({
       where: {
         id: id,
@@ -109,9 +128,38 @@ const deleteGenereById = async (req: Request, res: Response): Promise<void> => {
   }
 };
 
+const getGenresByCategory = async (
+  req: Request,
+  res: Response
+): Promise<void> => {
+  try {
+    const category: string = req.params.category;
+
+    const categoryId: number | null = await getCategoryIdByName(category);
+
+    if (!categoryId) {
+      res.json({ success: 0, error: "No category found" });
+      return;
+    }
+
+    const genres: Genre[] = await db.Genre.findAll({
+      attributes: ["genre"],
+      where: {
+        categoryId: categoryId,
+      },
+    });
+
+    res.json({ success: 1, result: genres });
+  } catch (error) {
+    logger.error(error);
+    res.json({ success: 0 });
+  }
+};
+
 export {
-  createGenreController,
-  getAllGenreController,
-  editGenreByIdController,
+  createGenre,
+  getAllGenre,
+  editGenreById,
   deleteGenereById,
+  getGenresByCategory,
 };
